+++ testing/black_box/config.json
@@ -1,0 +1,4 @@
+{"bin_parallel": "/Users/wr/Documents/Repos/Forked/tRIBS/cmake-build-parallel/tRIBSpar",
+  "bin_serial": "/Users/wr/Documents/Repos/Forked/tRIBS/cmake-build-serial/tRIBS",
+  "hj_path": "/Users/wr/Desktop/happy_jack",
+  "bs_path": "/Users/wr/Desktop/big_spring"}
